Make sycamore compile without default features
